Change pwa provider
