+++ tekblogg-web/.gitignore
@@ -47,3 +47,4 @@ sitemap-0.xml
 **/public/sw.js
 **/public/worker-*.js
 **/public/~partytown/
+**/public/swe-*.js
