+++ tekblogg-web/src/app/sw.ts
@@ -1,0 +1,16 @@
+import { defaultCache } from '@serwist/next/browser'
+import type { PrecacheEntry } from '@serwist/precaching'
+import { installSerwist } from '@serwist/sw'
+
+declare const self: ServiceWorkerGlobalScopeEventMap & {
+  // Change this attribute's name to your `injectionPoint`.
+  __SW_MANIFEST: (PrecacheEntry | string)[] | undefined
+}
+
+installSerwist({
+  precacheEntries: self.__SW_MANIFEST,
+  skipWaiting: true,
+  clientsClaim: true,
+  navigationPreload: true,
+  runtimeCaching: defaultCache
+})
